Feature/0.3.2/15 consultar perfil

+++ Saca la Bici/Data/Models/Objects/Profile.swift
@@ -20,5 +20,5 @@ struct Profile: Codable {
     let tiempoEnRecorrido: Double
     let rodadasCompletadas: Int
     let firebaseUID: String
-    let kilometrosMes: Int
+    let kilometrosMes: Double
 }
